Generalize error layout
This commit generalises the error layout from 404 in order to reuse it
for all other error types. It also moves the React error pages to
their own folder inside main; it seems much more nice and tidy that way.
