+++ lib/MusicBrainz/Server/Controller/Root.pm
@@ -166,7 +166,10 @@ sub error_404 : Private {
 
     $c->response->status(404);
     $c->stash->{current_view} = 'Node';
-    $c->stash->{component_props}{message} = $message;
+    $c->stash->{component_props} = {
+        message => $message,
+    };
+    $c->stash->{component_path} = 'main/error/404';
 }
 
 sub error_500 : Private
