+++ root/server/components.js
@@ -101,8 +101,8 @@ module.exports = {
   'label/LabelMerge': require('../label/LabelMerge'),
   'label/LabelRelationships': require('../label/LabelRelationships'),
   'label/SpecialPurpose': require('../label/SpecialPurpose'),
-  'main/404': require('../main/404'),
   'main/ConfirmSeed': require('../main/ConfirmSeed'),
+  'main/error/404': require('../main/error/404'),
   'main/index': require('../main/index'),
   'mbid/NotFound': require('../mbid/NotFound'),
   'oauth2/OAuth2Authorize': require('../oauth2/OAuth2Authorize'),
